Re-arranged Skill Class & Interface to have it read the Ability type before the Skill type. Skill type will take in a number and cast it as the SkillType's index.
Created AbilityType enum. Kitty has one in the  Manager script so need to talk to her about it.
Added all the Skill Names and types to use in the future.
The Skill DB Dictionary will take in the skill ID as it's key.

The Mob & Skill Database list is now private but can be seen in the inspector.
Cleaned up the scene so that the Utility and Databases are in it's on GameObject.